Add useOutlineEditor hook
