+++ packages/outline-playground/craco.config.js
@@ -15,6 +15,7 @@ module.exports = {
       'outline/HistoryHelpers': 'outline/dist/OutlineHistoryHelpers',
       'outline/KeyHelpers': 'outline/dist/OutlineKeyHelpers',
       // Outline React
+      'outline-react/useOutlineEditor': 'outline-react/dist/useOutlineEditor',
       'outline-react/useOutlineRichText':
         'outline-react/dist/useOutlineRichText',
       'outline-react/useOutlinePlainText':
