+++ packages/outline-playground/src/Editor.js
@@ -1,13 +1,20 @@
-// @flow strict-local
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict-local
+ */
 
 import type {OutlineEditor, ViewModel} from 'outline';
 
 import * as React from 'react';
-import {useCallback, useEffect, useMemo} from 'react';
-import {createEditor} from 'outline';
+import {useEffect, useMemo} from 'react';
 import useOutlineRichText from 'outline-react/useOutlineRichText';
 import useEmojis from './useEmojis';
 import useMentions from './useMentions';
+import useOutlineEditor from 'outline-react/useOutlineEditor';
 import usePlainText from 'outline-react/useOutlinePlainText';
 import useOutlineAutoFormatter from 'outline-react/useOutlineAutoFormatter';
 import useToolbar from './useToolbar';
@@ -26,66 +33,42 @@ const editorStyle = {
 
 type Props = {
   onChange: (ViewModel | null) => void,
+  onError: (Error) => void,
   isReadOnly?: boolean,
   isCharLimit?: boolean,
   isAutocomplete?: boolean,
 };
 
-function useOutlineEditor(
-  placeholder: string,
-): [OutlineEditor, (null | HTMLElement) => void] {
-  const editor = useMemo(
-    () =>
-      createEditor({
-        placeholder: 'editor-placeholder',
-        paragraph: 'editor-paragraph',
-        quote: 'editor-quote',
-        heading: {
-          h1: 'editor-heading-h1',
-          h2: 'editor-heading-h2',
-          h3: 'editor-heading-h3',
-          h4: 'editor-heading-h4',
-          h5: 'editor-heading-h5',
-        },
-        list: {
-          ol: 'editor-list-ol',
-          ul: 'editor-list-ul',
-        },
-        listitem: 'editor-listitem',
-        image: 'editor-image',
-        text: {
-          bold: 'editor-text-bold',
-          link: 'editor-text-link',
-          italic: 'editor-text-italic',
-          overflowed: 'editor-text-overflowed',
-          hashtag: 'editor-text-hashtag',
-          underline: 'editor-text-underline',
-          strikethrough: 'editor-text-strikethrough',
-          underlineStrikethrough: 'editor-text-underlineStrikethrough',
-          code: 'editor-text-code',
-        },
-        code: 'editor-code',
-      }),
-    [],
-  );
-
-  const editorElementRef = useCallback(
-    (editorElement: null | HTMLElement) => {
-      // Clear editorElement if not done already
-      if (editorElement !== null && editorElement.firstChild !== null) {
-        editorElement.textContent = '';
-      }
-      editor.setEditorElement(editorElement);
-    },
-    [editor],
-  );
-
-  useEffect(() => {
-    editor.setPlaceholder(placeholder);
-  }, [editor, placeholder]);
-
-  return [editor, editorElementRef];
-}
+const editorThemeClasses = {
+  placeholder: 'editor-placeholder',
+  paragraph: 'editor-paragraph',
+  quote: 'editor-quote',
+  heading: {
+    h1: 'editor-heading-h1',
+    h2: 'editor-heading-h2',
+    h3: 'editor-heading-h3',
+    h4: 'editor-heading-h4',
+    h5: 'editor-heading-h5',
+  },
+  list: {
+    ol: 'editor-list-ol',
+    ul: 'editor-list-ul',
+  },
+  listitem: 'editor-listitem',
+  image: 'editor-image',
+  text: {
+    bold: 'editor-text-bold',
+    link: 'editor-text-link',
+    italic: 'editor-text-italic',
+    overflowed: 'editor-text-overflowed',
+    hashtag: 'editor-text-hashtag',
+    underline: 'editor-text-underline',
+    strikethrough: 'editor-text-strikethrough',
+    underlineStrikethrough: 'editor-text-underlineStrikethrough',
+    code: 'editor-text-code',
+  },
+  code: 'editor-code',
+};
 
 function useOutlineOnChange(
   editor: OutlineEditor,
@@ -104,14 +87,6 @@ function useOutlineOnChange(
       return editor.addUpdateListener(onChange);
     }
   }, [onChange, editor]);
-  // Subscribe to errors
-  useEffect(() => {
-    if (editor !== null) {
-      return editor.addErrorListener((e) => {
-        throw e;
-      });
-    }
-  }, [editor]);
 }
 
 function ContentEditable({
@@ -139,12 +114,15 @@ function ContentEditable({
 
 export const useRichTextEditor = ({
   onChange,
+  onError,
   isReadOnly,
   isCharLimit,
   isAutocomplete,
 }: Props): [OutlineEditor, React.MixedElement] => {
   const [editor, editorElementRef] = useOutlineEditor(
     'Enter some rich text...',
+    onError,
+    editorThemeClasses,
   );
   const mentionsTypeahead = useMentions(editor);
   const props = useOutlineRichText(editor, isReadOnly);
@@ -185,12 +163,15 @@ export const useRichTextEditor = ({
 
 export const usePlainTextEditor = ({
   onChange,
+  onError,
   isReadOnly,
   isCharLimit,
   isAutocomplete,
 }: Props): [OutlineEditor, React.MixedElement] => {
   const [editor, editorElementRef] = useOutlineEditor(
     'Enter some plain text...',
+    onError,
+    editorThemeClasses,
   );
   const mentionsTypeahead = useMentions(editor);
   const props = usePlainText(editor, isReadOnly);
